added substr, updated .h and makefile

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: ereina-l <ereina-l@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/11 19:25:56 by ereina-l          #+#    #+#             */
-/*   Updated: 2024/09/27 11:35:20 by ereina-l         ###   ########.fr       */
+/*   Updated: 2024/09/27 13:17:35 by ereina-l         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
 	int	i;
